+++ src/java/org/apache/cassandra/db/compaction/Scrubber.java
@@ -44,7 +44,7 @@ public class Scrubber implements Closeable
 
     private final CompactionController controller;
     private final boolean isCommutative;
-    private final int expectedBloomFilterSize;
+    private final long expectedBloomFilterSize;
 
     private final RandomAccessReader dataFile;
     private final RandomAccessReader indexFile;
@@ -111,7 +111,7 @@ public class Scrubber implements Closeable
 
         this.expectedBloomFilterSize = Math.max(
             cfs.metadata.getMinIndexInterval(),
-            hasIndexFile ? (int)(SSTableReader.getApproximateKeyCount(toScrub)) : 0);
+            hasIndexFile ? SSTableReader.getApproximateKeyCount(toScrub) : 0);
 
         // loop through each row, deserializing to check for damage.
         // we'll also loop through the index at the same time, using the position from the index to recover if the
